Lots of fixes.
Client may only log in once as of now. This is because we have to
send EOF to not get Client stuck in an endless WAIT (not to be confused
with endless LOOP).

We have to send a killsignal of some sort after we've read the amount of
records we've wanted, maybe send a false boolean after objects?